[Server] move: 컨트롤러 분리 및 모듈화
