+++ server/models/post.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
     static associate(models) {
       // define association here
       post.belongsTo(models.user, { foreignKey: 'userId' });
-      post.hasMany(models.categories, { foreignKey: 'categoryId' });
+      post.hasMany(models.categories, { foreignKey: 'id' });
       post.hasOne(models.comments, { foreignKey: 'postId' });
       post.hasOne(models.user_place_likes, { foreignKey: 'postId' });
     }
